requirements.txt
pyfiglet
colorama
google-auth
google-auth-oauthlib
google-api-python-client
requests
moviepy==1.0.3
openai
python-dotenv
pygame
pillow
gTTS
pydub